Separate tpch tests into new file [Run CI]
